Create methodology page and wrapper
