+++ frontend/src/sites/cclw/pages/methodology.tsx
@@ -1,0 +1,12 @@
+import { ExternalLink } from "@components/ExternalLink";
+
+const Methodology = () => {
+  return (
+    <section>
+      <div className="text-content px-4 container mb-12">
+        <h1 className="my-8">Methodology</h1>
+      </div>
+    </section>
+  );
+};
+export default Methodology;
